client/requirements.txt
requests==2.25.1
pycryptodome==3.10.1
ecdsa==0.16.1
argparse==1.4.0
python-socketio==5.3.0
